requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn